// src/lib/api/schemas/customerAddressWithAddressRow.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { AddressId } from './addressId.ts';
import type { CustomerAddressId } from './customerAddressId.ts';
import type { CustomerAddressStatus } from './customerAddressStatus.ts';
import type { CustomerId } from './customerId.ts';

export interface CustomerAddressWithAddressRow {
  address_created_at: string;
  address_id: AddressId;
  city: string;
  contact_name: string;
  country: string;
  created_at: string;
  customer_id: CustomerId;
  id: CustomerAddressId;
  is_default_billing: boolean;
  is_default_shipping: boolean;
  phone: string;
  postal_code: string;
  state: string;
  status: CustomerAddressStatus;
  street_address: string;
  updated_at: string;
}